put play controls in separate viewModel
