+++ app/src/main/res/drawable/ic_times_all.xml
@@ -4,9 +4,9 @@
     android:tint="?colorOnSurface"
     android:viewportWidth="6.35"
     android:viewportHeight="6.35">
-    <path
-        android:fillAlpha="0.978"
-        android:fillColor="#535353"
-        android:pathData="m2.767,2.02c-0.03398,0.0708 -0.0665,0.14248 -0.09766,0.21517 -0.03115,0.0717 -0.06268,0.14739 -0.09477,0.22669 -0.03115,0.0793 -0.06334,0.16325 -0.09638,0.25198 -0.03304,0.0878 -0.06854,0.18369 -0.1063,0.28753h0.14312l0.08933,-0.2507h0.44442c0.0151,0.0425 0.0297,0.0846 0.04387,0.12615 0.01416,0.0406 0.02942,0.0821 0.04547,0.12455h0.15017c-0.03776,-0.10384 -0.07326,-0.19973 -0.1063,-0.28753 -0.03304,-0.0887 -0.06556,-0.17269 -0.09766,-0.25198 -0.03115,-0.0793 -0.06234,-0.15496 -0.09349,-0.22669 -0.03115,-0.0727 -0.06367,-0.14437 -0.09766,-0.21517zM2.82656,2.17722c0.03209,0.0708 0.06263,0.14481 0.09189,0.22221 0.03021,0.0765 0.06041,0.15577 0.09061,0.2379h-0.36533c0.02926,-0.0821 0.05879,-0.16144 0.08805,-0.2379 0.03021,-0.0774 0.06174,-0.15141 0.09478,-0.22221zM3.82966,2.49356v0.61668h-3.3002v0.60195h3.2998v0.617l1.9916,-0.91797z"
-        android:strokeAlpha="0.978" />
+  <path
+      android:fillAlpha="0.978"
+      android:fillColor="#FF000000"
+      android:pathData="m2.1118,1.73001c-0.10011,0.20856 -0.19607,0.41995 -0.28784,0.63407 -0.09177,0.21134 -0.18451,0.43356 -0.27905,0.66714 -0.04509,0.11478 -0.09309,0.24335 -0.13953,0.36484h0.39998c0.06745,-0.18646 0.13512,-0.37204 0.20257,-0.54829 0.08899,-0.22802 0.18172,-0.44618 0.27905,-0.65474 0.09455,0.20856 0.1851,0.42672 0.2713,0.65474 0.06964,0.17627 0.13965,0.3618 0.20929,0.54829h0.42013c-0.04713,-0.1215 -0.09513,-0.25005 -0.14159,-0.36484 -0.09177,-0.23358 -0.18315,-0.4558 -0.27492,-0.66714 -0.09177,-0.21412 -0.18773,-0.42551 -0.28784,-0.63407zM3.1877,3.396c0.04878,0.12576 0.09674,0.2448 0.14624,0.37775 0.02197,0.0584 0.04811,0.13041 0.0708,0.19121h0.51676v0.58291l2.2727,-0.86713 -2.2727,-0.86713v0.58239zM3.40474,3.96497h-0.43357c0.03409,0.0968 0.06784,0.19328 0.10025,0.28835 0.04171,0.11958 0.08605,0.24229 0.13333,0.36742h0.44235c-0.08612,-0.23682 -0.16458,-0.44729 -0.24236,-0.65577zM2.97117,3.96497c-0.00978,-0.0278 -0.01902,-0.0548 -0.02894,-0.0827h-1.31l-0.02946,0.0827zM1.60277,3.96497h-0.41238c-0.07778,0.20848 -0.15624,0.41895 -0.24236,0.65577h0.42168zM1.19039,3.96497c0.02268,-0.0608 0.04882,-0.13281 0.0708,-0.19121 0.0495,-0.13295 0.09611,-0.25199 0.14418,-0.37775h-1.2495v0.56896zM2.76759,3.396h-0.96222c-0.01875,0.0518 -0.03757,0.0998 -0.05633,0.15245h1.0764c-0.01935,-0.0526 -0.03853,-0.10063 -0.05788,-0.15245z"
+      android:strokeAlpha="0.978" />
 </vector>
