+++ app/src/main/res/drawable/ic_times_one.xml
@@ -4,9 +4,9 @@
     android:tint="?colorOnSurface"
     android:viewportWidth="6.35"
     android:viewportHeight="6.35">
-    <path
-        android:fillAlpha="0.978"
-        android:fillColor="#1a1a1a"
-        android:pathData="m2.78899,2.02c-0.04247,0.0491 -0.08927,0.0902 -0.14024,0.12327 -0.05098,0.0321 -0.1034,0.0586 -0.15721,0.0794l0.04131,0.0989c0.01983,-0.006 0.04007,-0.0146 0.06083,-0.024 0.02172,-0.009 0.04196,-0.0194 0.06083,-0.0297 0.01983,-0.0103 0.03775,-0.0209 0.05379,-0.0324 0.01699,-0.0113 0.03125,-0.0223 0.04259,-0.0326v0.79854h0.1316v-0.98137zM3.82899,2.49356v0.61668h-3.3002v0.60195h3.2998v0.617l1.9916,-0.91797z"
-        android:strokeAlpha="0.978" />
+  <path
+      android:fillAlpha="0.978"
+      android:fillColor="#FF000000"
+      android:pathData="m2.681,1.79001c-0.13514,0.13812 -0.2838,0.25378 -0.44597,0.34675 -0.16217,0.0903 -0.32905,0.1648 -0.50023,0.22324l0.13074,0.27853c0.06307,-0.0186 0.12772,-0.0411 0.19379,-0.0677 0.06907,-0.0266 0.13372,-0.0545 0.19379,-0.0837 0.06307,-0.0292 0.12,-0.0596 0.17105,-0.0915 0.05406,-0.0319 0.09935,-0.0623 0.13539,-0.0915v1.0831h0.41858v-1.5973zM2.97814,3.3873v0.58032h0.93379v0.59428l2.2448,-0.8847 -2.2443,-0.8847v0.5948zM2.97814,3.96762h-0.41858v0.58343h0.41858zM2.55956,3.96762v-0.58032h-2.3663v0.58032z"
+      android:strokeAlpha="0.978" />
 </vector>
