+++ app/src/main/res/layout/partial_bottom_play_controls.xml
@@ -6,12 +6,13 @@
     <data>
 
         <variable
-            name="viewModel"
-            type="com.techbeloved.hymnbook.hymndetail.HymnPagerViewModel" />
+            name="nowPlaying"
+            type="com.techbeloved.hymnbook.nowplaying.NowPlayingViewModel" />
 
     </data>
 
     <androidx.cardview.widget.CardView
+        android:id="@+id/cardView_play_controls"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:clickable="true"
@@ -112,7 +113,7 @@
                 android:clickable="true"
                 android:contentDescription="@string/play_pause_description"
                 android:focusable="true"
-                android:src="@{viewModel.isPlaying ? @drawable/ic_pause : @drawable/ic_play}"
+                android:src="@{nowPlaying.isPlaying ? @drawable/ic_pause : @drawable/ic_play}"
                 app:layout_constraintBottom_toBottomOf="@+id/imageView_controls_previous"
                 app:layout_constraintEnd_toStartOf="@+id/imageView_controls_next"
                 app:layout_constraintHorizontal_bias="0.5"
